requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
attrs==23.1.0
black==23.3.0
browsermob-proxy==0.8.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
clipboard==0.0.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
filelock==3.11.0
flake8==6.0.0
fonttools==4.39.3
identify==2.5.22
idna==3.4
imageio==2.27.0
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
lxml==4.9.2
matplotlib==3.7.1
mccabe==0.7.0
minio==7.1.14
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.7.0
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
pika==1.3.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyMuPDF==1.21.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-interface==1.6.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
selenium==3.141.0
six==1.16.0
testit-adapter-pytest==2.1.5
testit-api-client==3.0.1
testit-importer-allure==1.3.0
testit-python-commons==2.1.5
tifffile==2023.3.21
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
xmltodict==0.13.0
zipp==3.15.0
